Updated and fixed permissions checkings

+++ achievements_ms/resolvers/achievements.resolver.js
@@ -1,9 +1,12 @@
 const { GraphQLError } = require('graphql');
 const checkHabitOwnership = require('../../utils/checkHabitOwnership');
+const validatePermissions = require('./../../utils/matchPermisssion');
 
 const resolvers = {
     Query: {
-        achievementsByHabit: async (_, { id, page, per_page }, { dataSources, userId, isAdmin }) => {
+        achievementsByHabit: async (_, { id, page, per_page }, { dataSources, userId, isAdmin, permissions, role }) => {
+
+            validatePermissions(permissions, role, "achievementsByHabit");
 
             // Check first if user is allowed to access this habit
             const allowed = await checkHabitOwnership(dataSources.habitsAPI, userId, isAdmin, id);
@@ -16,7 +19,10 @@ const resolvers = {
     },
 
     Mutation: {
-        addAchievement: async (_, { achievement }, { dataSources, userId, isAdmin }) => {
+        addAchievement: async (_, { achievement }, { dataSources, userId, isAdmin, permissions, role }) => {
+
+            validatePermissions(permissions, role, "addAchievement");
+
             // Check first if user is allowed to access this habit
             const allowed = await checkHabitOwnership(dataSources.habitsAPI, userId, isAdmin, achievement.habId);
 
@@ -26,11 +32,13 @@ const resolvers = {
             return dataSources.achievementsAPI.addAchievement(achievement);
         },
 
-        deleteAchievement: async (_, { id }, { dataSources }) => {
+        deleteAchievement: async (_, { id }, { dataSources, permissions, role }) => {
+            validatePermissions(permissions, role, "deleteAchievement");
             return dataSources.achievementsAPI.deleteAchievement(id);
         },
 
-        updateAchievement: async (_, { achievement }, { dataSources, userId, isAdmin }) => {
+        updateAchievement: async (_, { achievement }, { dataSources, userId, isAdmin, permissions, role }) => {
+            validatePermissions(permissions, role, "updateAchievement");
             if (achievement.habId) {
                 // Check first if user is allowed to access this habit
                 const allowed = await checkHabitOwnership(dataSources.habitsAPI, userId, isAdmin, achievement.habId);
@@ -43,11 +51,13 @@ const resolvers = {
             return dataSources.achievementsAPI.updateAchievement(id, achievement);
         },
 
-        updateStreak: async (_, { id, retainStreak }, { dataSources }) => {
+        updateStreak: async (_, { id, retainStreak }, { dataSources, permissions, role }) => {
+            validatePermissions(permissions, role, "updateStreak");
             return dataSources.achievementsAPI.updateAchievementStreak(id, retainStreak);
         },
 
-        notifyStreakUpdate: async (_, { hab_id, freq_type, streak }, { dataSources }) => {
+        notifyStreakUpdate: async (_, { hab_id, freq_type, streak }, { dataSources, permissions, role }) => {
+            validatePermissions(permissions, role, "notifyStreakUpdate");
             return dataSources.achievementsAPI.notifyStreakUpdate(hab_id, freq_type, streak);
         }
